+++ src/init.sh
@@ -22,6 +22,8 @@ useremail=$(git config --global user.email || echo 'noreply@yoda.org')
 
 touch .dockerignore
 mkdir -p $yoda_dir/{images,containers,.ssh/connections}
+sed 's/{{driver}}/host/g' "$YODA_PATH/templates/networks.yml" > "$yoda_dir/containers/networks.yml"
+sed 's/{{driver}}/bridge/g' "$YODA_PATH/templates/networks.yml" > "$yoda_dir/containers/networks.dev.yml"
 cp $YODA_PATH/templates/env.sh $yoda_dir
 cp $YODA_PATH/templates/{Env,Build,Start}file $yoda_dir
 sed "s/{{user}}/$username/g;s/{{email}}/$useremail/g;" $YODA_PATH/templates/Dockerfile > $yoda_dir/images/Dockerfile-base